+++ core/src/main/java/org/vertexium/DefaultGraphVisitor.java
@@ -1,0 +1,33 @@
+package org.vertexium;
+
+public class DefaultGraphVisitor implements GraphVisitor {
+    @Override
+    public void visitElement(Element element) {
+
+    }
+
+    @Override
+    public void visitVertex(Vertex vertex) {
+
+    }
+
+    @Override
+    public void visitEdge(Edge edge) {
+
+    }
+
+    @Override
+    public void visitProperty(Element element, Property property) {
+
+    }
+
+    @Override
+    public void visitExtendedDataRow(Element element, String tableName, ExtendedDataRow row) {
+
+    }
+
+    @Override
+    public void visitProperty(Element element, String tableName, ExtendedDataRow row, Property property) {
+
+    }
+}